+++ NEWS.md
@@ -1,3 +1,15 @@
+## 0.1.5
+### New Features
+- Introduced application theme configuration file generator RStudio add-in
+
+### Enhancements
+- Allowed header to have navbar menu
+- Renamed `createAlert` method to be `createPSAlert` to:
+  - Avoid confusion with `bs4Dash::createAlert` method
+  - Display alert with status icon correctly when there is no alert title
+
+-----
+
 ## 0.1.4
 ### New Features
 - Introduced Announcements configuration file generator RStudio add-in